% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{diffn}
\alias{diffn}
\title{Finite differences}
\usage{
diffn(m, ndiff = 1L)
}
\arguments{
\item{m}{2- or more-column numeric matrix}

\item{ndiff}{Integer scalar, order of finite difference (1 by default)}
}
\value{
Numeric matrix, first column is midpoints of x, the second
 and following are dy/dx
}
\description{
Calculate dy/dx where x,y are first and the rest of columns in the entry matrix 'm'
}
