% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{dmat}
\alias{dmat}
\title{Differentiation matrix}
\usage{
dmat(f, nqw = NULL, xk = NULL, n = NULL)
}
\arguments{
\item{nqw}{Integer scalar, row number of qw matrix (i.e. degree of freedom of a B-spline)}

\item{xk}{Numeric vector, knot positions}

\item{n}{Integer scalar, B-spline polynomial order (3 by deafult)}
}
\value{
Numeric matrix of size \code{nqw-1 x nqw}
}
\description{

}
