% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{fit_spline}
\alias{fit_spline}
\title{Fitting regression spline estimator}
\usage{
fit_spline(x_values, y_values, interior_knots, degree)
}
\arguments{
\item{x_values, y_values}{Numeric vectors}

\item{interior_knots}{Numeric vector representing interior knots sequence. It can be obtained from \code{\link{knots_quantile}} function based on the given x points.}

\item{degree}{Numeric value indicating the spline degree}
}
\value{
List containing the spline coefficients estimates, knots sequence, spline degree
}
\description{
This function computes the coefficients spline regression estimator.
}
\examples{
set.seed(923)
n = 30
x_values = sort(runif(n, 0, 1))
y_values = sin(2 * pi * x_values) + cos(4 * pi * x_values) + rnorm(n, sd = 0.2)
#spline degree specification
degree = 3
# knot generation
num_interior_knots = 5
interior_knots = knots_quantile(x_values, num_interior_knots)
# model fitting
model = fit_spline(x_values, y_values, interior_knots, degree)
print(model)

}
