% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ipk}
\alias{ipk}
\title{Intervals of points in knot intervals}
\usage{
ipk(x, xk)
}
\arguments{
\item{x}{Numeric vector, abscissa points (must be non decreasing)}

\item{xk}{Numeric vector, knots (must be non decreasing)}
}
\value{
Integer matrix of size \code{(2 x length(xk)-1)}. Indexes are 0-based
}
\description{
Find first and last+1 indexes iip s.t. x[iip] belongs to interval starting at xk[iik]
}
