% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jacw}
\alias{jacw}
\title{Knot Jacobian of B-spline with weights}
\usage{
jacw(jac, qws)
}
\arguments{
\item{jac}{Numeric array, such as returned by \code{bsc(..., cjac=TRUE)}}

\item{qws}{Numeric matrix, each column is a set of weights forming a
B-spline. If qws is a vector, it is coerced to 1-column matrix.}
}
\value{
Numeric array of size \code{nx x ncol(qw) x nk}, where \code{nx=dim(jac)[1]}
  and nk is the number of knots \code{dim(jac)[3]+n+1} (n being polynomial order).
}
\description{
Knot Jacobian of B-spline with weights
}
