% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{knots_quantile}
\alias{knots_quantile}
\title{Generating interior knots squence for given data points}
\usage{
knots_quantile(x, dimension, degree = 3)
}
\arguments{
\item{x}{Numeric vector representing data points}

\item{dimension}{Numeric value indicating the number of basis functions.}

\item{degree}{Numeric value representing the spline degree. The default value is set to 3.}
}
\value{
Numeric vector with interior knots sequence
}
\description{
This function .....
}
\examples{
x = runif(50, 0, 1)
knots = knots_quantile(x, 10, 3)
print(knots)
}
