% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{plot_spline}
\alias{plot_spline}
\title{Plotting spline estimator with scatter plots}
\usage{
plot_spline(x_values, y_values, model, grid_x)
}
\arguments{
\item{x_values, y_values}{Numeric vector}

\item{model}{List object obtained from the \code{\link{fit_spline}} function}

\item{grid_x}{Numeric vector with a grid of evaluation points}
}
\description{
This function provides a plot of the spline estimator with data points
}
