% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{predict_spline}
\alias{predict_spline}
\title{Predicting the values of y for given x based on the spline estimator}
\usage{
predict_spline(model, new_x)
}
\arguments{
\item{model}{List object obtained from the \code{\link{fit_spline}} function}

\item{new_x}{Numeric vector representing a grid of evaluation points}
}
\value{
Numeric vector with predicted values at new_x
}
\description{
This function computes the predicted values of y for given x based on the spline estimator.
}
