% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jacw}
\alias{jacw}
\title{Internal knot Jacobian of B-spline with weights}
\usage{
jacw(jac, qw)
}
\arguments{
\item{jac}{Numeric array, such as returned by \code{bsc(..., cjac=TRUE)}}

\item{qw}{Numeric matrix, each column is a set of weights forming a B-spline}
}
\value{
Numeric array of size nx x ncol(qw) x nki, where \code{nx=dim(jac)[1]}
  and nki is the number of internal knots \code{dim(jac)[3]-n-1} (n being the
  polynomial order of B-spline)
}
\description{
Internal knot Jacobian of B-spline with weights
}
